Meshes Cleanup and Voxel Models
Removed old redundant mesh types (SquareMesh and CubeMesh) and replaced them with static functions in Mesh.

ModelMesh inherits from IndicesMesh

Meshes no longer have a shape field, allowing for ArrayMesh and IndicesMesh to be instantiated without an anonymous inheriting class. Instead there is a getBounds method which uses the VertexAttributes to get the min and max positions of the vertices, a transform, can be applied to the vertices first if required.

MeshVertex and IndicesMeshBuilder can be used to more easily generate meshes.

Added VoxModelLoader which is able to import .vox files as models. It uses greedy meshing to generate the mesh.