Framebuffer improvements and Editor bug fixing.
Cleaned up Textures. They now store a single id String instead of a fileLocation and fileName. There is also a static create function to make a blank texture with a size and format.

Added a copy to texture function to Framebuffer. This allows ObjectRenderer to create multiple textures, and for those textures to persist after deleting the renderer.

Fixed a bunch of bugs that had built up in the Editor.