New Dropdown Menu, added Shape2D and GameObject field editors to the Editor.
Added DropdownComponent, which has a static function to create a dropdown components made of a text display for the selected option, SelectableScrollList of options and a button to open and close the list.

The Shape2D field editor uses the new dropdown menu to select the shape type, currently either Rectangle or Parallelogram, and then generates fields needed to edit the chosen shape. These fields create a new shape and set the shape field instead of modifying the existing shape.

The GameObject field stores the object in relation to the parent object of the component, and so needs the context of the component it is part of and should wait until the object has been parsed and added to an ObjectCollection. To do this ComponentField#isLateParse has been added to tell the parser if the field and it's value should be stored and parsed at the end.

Moved the transform functions from MenuItem to GameObject, and added an optional layer argument to GameObject's constructor. This is to remove the need for MenuItem, to in the future remove all inheritors of GameObject except Prefab, as all functionality should be stored in Components.